# performance: optimising purego performance by expanding loops