@aimuz 2024-01-09
Maybe assembly can be optimised this way too